requirements.txt
asgiref==3.4.1
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.4
Django==4.0.1
django-appconf==1.0.5
django-ckeditor==6.2.0
django-crispy-forms==1.14.0
django-environ==0.8.1
django-filter==21.1
django-import-export==2.7.1
django-js-asset==1.2.2
django-staticfiles==1.2.1
djangorestframework==3.13.1
et-xmlfile==1.1.0
filelock==3.4.2
gunicorn==20.1.0
Markdown==3.3.6
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.9
Pillow==8.4.0
platformdirs==2.4.1
python-decouple==3.6
pytz==2021.3
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
sweetify==2.3.0
tablib==3.2.0
tzdata==2021.5
virtualenv==20.13.0
xlrd==2.0.1
xlwt==1.3.0
django-phonenumber-field
